requirements.txt
python-dotenv==1.0.1
pydantic==2.8.2
openai>=1.40.0
requests>=2.31.0
beautifulsoup4>=4.12.3
lxml==5.2.2
jsonschema==4.22.0
validators>=0.20.0
PyYAML>=6.0.1
